client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "mcp[cli]>=1.9.3",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
]
